requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
httpx
google-genai
python-dotenv
sentence-transformers
sqlalchemy
pymysql
passlib[bcrypt]
python-jose[cryptography]
email-validator
pdfplumber>=0.11.9
sentence-transformers>=5.3.0
